test(workflow): split task-repair-audit-resilience test to satisfy FR-006
The merged test file exceeded the 500-line new-file cap (591 lines),
which the quality CI job's check-test-file-cap ratchet fails closed on.
Split by functional area, following the repo's test-file-split protocol:

- task-repair-audit-resilience-2098.test.ts (284 lines): the core
  ensureAuditEvent/recoverPreparedTaskRepair resilience tests (FB-001).
- task-repair-wal-limits-2098.test.ts (239 lines): corrupt-WAL error
  messaging (FB-002), force-repair field length limits (FB-004), and
  the JSON-escaping safety tests (FB-001).
- tests/helpers/task-repair-2098-fixtures.ts: shared TASK_ID/repairArgs/
  seedSettledRepairState fixtures extracted to avoid duplicating them
  across both split files.

No behavioral change; isolated and co-run verified identical (10 tests,
35 expect() calls, same as the unsplit file).
